refactor: DTO validation 추가 및 폴더 구조 정리
