+++ src/main/java/com/sparta/todayeats/store/dto/request/StoreHiddenRequest.java
@@ -1,4 +1,4 @@
-package com.sparta.todayeats.store.dto;
+package com.sparta.todayeats.store.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
