+++ src/main/java/com/sparta/todayeats/store/dto/response/StoreHiddenResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.todayeats.store.dto;
+package com.sparta.todayeats.store.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
