+++ src/main/java/com/sparta/todayeats/store/service/StoreService.java
@@ -6,7 +6,12 @@ import com.sparta.todayeats.category.domain.entity.Category;
 import com.sparta.todayeats.category.domain.repository.CategoryRepository;
 import com.sparta.todayeats.global.response.PageResponse;
 import com.sparta.todayeats.global.exception.*;
-import com.sparta.todayeats.store.dto.*;
+import com.sparta.todayeats.store.dto.request.StoreCreateRequest;
+import com.sparta.todayeats.store.dto.request.StoreHiddenRequest;
+import com.sparta.todayeats.store.dto.request.StoreUpdateRequest;
+import com.sparta.todayeats.store.dto.response.StoreCreateResponse;
+import com.sparta.todayeats.store.dto.response.StoreHiddenResponse;
+import com.sparta.todayeats.store.dto.response.StoreResponse;
 import com.sparta.todayeats.store.entity.Store;
 import com.sparta.todayeats.store.repository.StoreRepository;
 import com.sparta.todayeats.user.domain.entity.User;
